Allow passing in custom alphabet
By default we use a checkerboard alphabet containing the 26 uppercase
ASCII characters, plus `.` and `/`.  This isn't always what's wanted, so
let users pass in a custom alphabet to use.

Currently it needs to be 28 characters long.  In the future we should
allow up to 100 characters by allowing users to specify a checkerboard
key containing an arbitrary number of spaces.

Signed-off-by: Johannes Löthberg <johannes@kyriasis.com>